Deployment part 1 - setup tslint and gulpfiles
