+++ client/package.json
@@ -4,6 +4,7 @@
   "private": true,
   "homepage": "https://hackgt.github.io/bolt",
   "dependencies": {
+    "@types/accounting": "^0.4.1",
     "@types/jest": "24.0.11",
     "@types/node": "11.11.3",
     "@types/react": "^16.8.8",
