+++ tslint.json
@@ -1,0 +1,76 @@
+{
+  "extends": "tslint:recommended",
+  "rules": {
+    "max-line-length": [false],
+    "no-unnecessary-initializer": false,
+    "no-var-requires": false,
+    "no-consecutive-blank-lines": false,
+    "quotemark": [
+      true,
+      "double",
+      "avoid-escape"
+    ],
+    "interface-name": [false],
+    "no-empty-interface": false,
+    "no-namespace": false,
+    "ordered-imports": false,
+    "object-literal-sort-keys": false,
+    "arrow-parens": false,
+    "no-string-literal": false,
+    "indent": {
+      "options": ["spaces", 4]
+    },
+    "member-ordering": [
+      true,
+      {
+        "order": [
+          "public-static-field",
+          "public-static-method",
+          "protected-static-field",
+          "protected-static-method",
+          "private-static-field",
+          "private-static-method",
+          "public-instance-field",
+          "protected-instance-field",
+          "private-instance-field",
+          "public-constructor",
+          "protected-constructor",
+          "private-constructor",
+          "public-instance-method",
+          "protected-instance-method",
+          "private-instance-method"
+        ]
+      }
+    ],
+    "no-console": [
+      true,
+      "debug",
+      "info",
+      "time",
+      "timeEnd",
+      "trace"
+    ],
+    "no-inferrable-types": [
+      true,
+      "ignore-params"
+    ],
+    "no-switch-case-fall-through": true,
+    "typedef": [
+      true,
+      "call-signature",
+      "parameter"
+    ],
+    "trailing-comma": [
+      true,
+      {
+        "multiline": {
+          "objects": "ignore",
+          "arrays": "ignore",
+          "functions": "never",
+          "typeLiterals": "ignore"
+        },
+        "singleline": "never"
+      }
+    ]
+  }
+}
